Improve PasswordField further
